+++ src/Exception/RequestException.php
@@ -45,9 +45,13 @@ class RequestException extends TransferException implements RequestExceptionInte
 
     /**
      * Wrap non-RequestExceptions with a RequestException
+     *
+     * @deprecated since 7.11. Create a RequestException or ResponseException directly instead.
      */
     public static function wrapException(RequestInterface $request, \Throwable $e): RequestException
     {
+        \trigger_deprecation('guzzlehttp/guzzle', '7.11', '%s::wrapException() is deprecated and will be removed in 8.0. Create a %s or %s directly instead.', self::class, self::class, ResponseException::class);
+
         return $e instanceof RequestException ? $e : new RequestException($e->getMessage(), $request, null, $e);
     }
 
@@ -86,7 +90,7 @@ class RequestException extends TransferException implements RequestExceptionInte
             $className = ServerException::class;
         } else {
             $label = 'Unsuccessful request';
-            $className = __CLASS__;
+            $className = ResponseException::class;
         }
 
         $uri = \GuzzleHttp\Psr7\Utils::redactUserInfo($request->getUri());
@@ -121,17 +125,26 @@ class RequestException extends TransferException implements RequestExceptionInte
 
     /**
      * Get the associated response
+     *
+     * Calling this method is deprecated since 7.11. Use instanceof
+     * ResponseException and ResponseException::getResponse() instead.
      */
     public function getResponse(): ?ResponseInterface
     {
+        \trigger_deprecation('guzzlehttp/guzzle', '7.11', '%s::getResponse() is deprecated and will be removed in 8.0. Use instanceof %s and %s::getResponse() instead.', static::class, ResponseException::class, ResponseException::class);
+
         return $this->response;
     }
 
     /**
      * Check if a response was received
+     *
+     * @deprecated since 7.11. Use instanceof ResponseException instead.
      */
     public function hasResponse(): bool
     {
+        \trigger_deprecation('guzzlehttp/guzzle', '7.11', '%s::hasResponse() is deprecated and will be removed in 8.0. Use instanceof %s instead.', static::class, ResponseException::class);
+
         return $this->response !== null;
     }